pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aware-parser-adk"
version = "0.1.0"
description = "Minimal ParserAgent (Google ADK) for BuildSpec extraction and validation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-adk>=1.0.0",
  "pydantic>=2.7,<3.0",
  "typer>=0.12,<1.0",
  "requests>=2.31,<3.0",
  "fastapi>=0.111,<1.0",
  "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
aware = "cli.main:run"
rca = "cli.main:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "models*", "aware_models*", "llm*", "runtime*", "cli*", "ui*", "tools*", "templates*"]

[tool.pytest.ini_options]
pythonpath = ["."]
